arpeggiated notes added

+++ utils_decoder.py
@@ -4,8 +4,7 @@ from utils_coder import FREQS
 import numpy as np
 import librosa
 import wave # for .wav format
-# import os
-# os.system("clear")
+
 
 #file = wave.open("mensaje.wav", "r") # rb = read binary
 def cargar_audio(ruta_archivo):
@@ -16,7 +15,7 @@ def cargar_audio(ruta_archivo):
     if len(datos.shape) == 2:
         datos = datos[:, 0]
 
-    audio = datos / np.max(np.abs(datos)) # normalizar los datos
+    audio = datos.astype(np.float32) / np.max(np.abs(datos)) # normalizar los datos
     
     return y, sr, audio
